move models/*Payload to a particular package
