+++ sdk-java8/src/main/java/ru/tinkoff/invest/openapi/okhttp/MarketContextImpl.java
@@ -6,7 +6,7 @@ import okhttp3.*;
 import org.jetbrains.annotations.NotNull;
 import ru.tinkoff.invest.openapi.MarketContext;
 import ru.tinkoff.invest.openapi.exceptions.OpenApiException;
-import ru.tinkoff.invest.openapi.models.RestResponse;
+import ru.tinkoff.invest.openapi.payload.RestResponse;
 import ru.tinkoff.invest.openapi.models.market.*;
 
 import java.io.IOException;
